ref: moved peering conformance tests
