+++ src/tests/conformance/mod.rs
@@ -2,4 +2,5 @@ mod basic_query;
 mod handshake;
 mod invalid_message;
 mod messages;
+mod peering;
 mod unsolicited_response;
